requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
PyMySQL==1.0.3
requests==2.31.0